third_party/WebKit/LayoutTests/fullscreen/full-screen-iframe-legacy-expected.txt

Test for bug 61461: Handle entering full screen security restrictions. Note that the original pass condition has been reversed.

To test manually, click the video "full screen" button - the page should NOT enter full screen mode.

EVENT(webkitfullscreenerror)
END OF TEST
